fix: fix 4 issues
