+++ src/components/Footer/Footer.types.ts
