--- src/components/FooterComponent/FooterComponentsStyles.tsx
@@ -1,7 +1,0 @@
-import { styled } from '@mui/system';
-
-export const FooterTeamWrapper = styled('div')({
-  width: '100%',
-  display: 'flex',
-  justifyContent: 'space-between',
-});
