--- src/components/FooterComponent/index.ts
@@ -1,1 +1,0 @@
-export { FooterComponent } from './FooterComponent';
