+++ src/constants/common.dictionary.ts
@@ -6,30 +6,3 @@ export const PATH = {
 export const MAIN_PAGE_TITLE = 'Main Page Title';
 export const WELCOME_PAGE_TITLE = 'Welcome to Lemasello';
 export const PAGE_404_TITLE = 'Error 404, this page can not be found...';
-
-export const TEAM = [
-  {
-    id: 1,
-    name: 'leonid',
-    link: 'https://github.com/LGers',
-    avatar: 'https://avatars.githubusercontent.com/u/87089817?v=4',
-  },
-  {
-    id: 2,
-    name: 'marina',
-    link: 'https://github.com/lukashkinamarina',
-    avatar: 'https://avatars.githubusercontent.com/u/86889005?v=4',
-  },
-  {
-    id: 3,
-    name: 'sergio',
-    link: 'https://github.com/sergioivanov008',
-    avatar: 'https://avatars.githubusercontent.com/u/74328346?s=400&v=4',
-  },
-];
-
-export const FOOTER_DATA = {
-  FOOTER_YEAR: '© 2022',
-  FOOTER_LINK: 'https://rs.school/js/',
-  FOOTER_LOGO: 'https://rs.school/images/rs_school_js.svg',
-};
