+++ src/pages/MainPage/MainPage.tsx
@@ -9,7 +9,7 @@ import {
 import { MAIN_PAGE_TITLE, PATH } from '../../constants/common.dictionary';
 import { WelcomePageHeader } from '../../components/WelcomePageHeader';
 import { Link } from 'react-router-dom';
-import { FooterComponent } from '../../components/FooterComponent';
+import { Footer } from '../../components/Footer';
 
 export const MainPage = () => {
   return (
@@ -24,7 +24,7 @@ export const MainPage = () => {
         </Content>
         <FooterWrapper>
           {/* <p>Footer</p> */}
-          <FooterComponent />
+          <Footer />
         </FooterWrapper>
       </Wrapper>
     </BodyWrapper>
