+++ src/pages/Page404/Page404.test.tsx
@@ -3,5 +3,5 @@ import { Page404 } from './Page404';
 
 test('renders Page404 component', () => {
   render(<Page404 />);
-  expect(screen.getByText(`Page404`)).toBeInTheDocument();
+  expect(screen.getByText(`Error Message`)).toBeInTheDocument();
 });
